qup_i2c: Reads greater than 256 bytes are processed.
QUP hardware does not allow reads more than 256 bytes to be sent to
I2C mini core without a STOP. This commit breaks such reads in chunks
of 256 bytes. It is assumed that the slave will not have problem with
the STOP per chunk sent by QUP hardware and that the slave would treat
it as 1 contiguous read without STOP(s) in between.

Change-Id: I0c6f0b9395ba6fcf1c4cee5ccefed9f5f948c3b7
Signed-off-by: Sagar Dharia <sdharia@codeaurora.org>