msm: audio: qdsp5v2: Update mvs driver to support back to back enable
The existing mvs driver wait for command acknowledgement
immediately and also it tries to open the control handling
thread on open. This results in failure when back to back
driver operation executed. Updated MVS driver to open control
handling thread on init and wait for command acknowledgement
on approriate place

Change-Id: Ic29925db5d921fac4702292e4a5b621006303eac
CRs-fixed: 270332
Signed-off-by: Neema Shetty <nshetty@codeaurora.org>
Signed-off-by: Rajesha Kini <rkini@codeaurora.org>