Revert "mmc: sdio: optimized SDIO IRQ handling for single irq"
This reverts commit 06e8935febe687e2a561707d4c7ca4245d261dbe.
With the optimization, the sdio_irq_thread() sends a spurious
interrupt notification to the registered function handler during
ksdioirq thread start up. The function handler may or may not
handle this interrupt gracefully. The correct way to inform
about interrupt is to check CCCR_INTx register for pending
interrupts before calling function driver's irq handler
which is broken with this optimization, hence the revert.

CRs-Fixed: 345547
Change-Id: Ia1525ef439c8d0d36b796f393066065648356635
Signed-off-by: Sujit Reddy Thumma <sthumma@codeaurora.org>